Switch to long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[] {4,40,25,600,15};
+        long[] sales = new long[] {4,40,25,600,15};
         SalesManager salesManager = new SalesManager(sales);
 
-         int max = salesManager.max();
+         long max = salesManager.max();
         System.out.println(max);
 
     }
